Redid how the parser works so that expressions inside function calls are evaluated before being passed to the function. What was happening was that a call to
    $length( "one" & "two" & "three" )

was generating

    erlang:length(["one", "two", "three"])

which is total nonsense. Instead the following is now generated:

    erlang:length("one" ++ "two" ++ "three")

And that is now done for all functions and list of things. I.e. expressions are
evaluated before being used.

Also implemente the '%' modulo operator which allows for doing things like:

    ($$.payload + 1) % $$.some_maximum

as a counter for frames for example.

And finally also improved length so that doing length on a binary value uses
erlang:byte_size/1 and not erlang:length/1 - which errored out.